try extract merge to tools
